refactor -> 불필요한 줄바꿈 제거, Game package 생성 후, Game 처리 관련된 클래스 관리
