+++ src/main/java/Game/Controller.java
@@ -1,4 +1,4 @@
-package baseball;
+package Game;
 
 import java.util.ArrayList;
 import java.util.List;
